WIP: TETRIS server and library
The TETRIS server is now able to manage clients. It can set the
affinity according to a (randomly) selected mapping for every thread
of a registered client.

The client library properly intercepts the pthread calls and forwards
the information to the TETRIS server which can then set the proper
affinity.