Make TERIS_DYNAMIC_CLIENT option to a multi-value option and rename
Before, TETRiS decided whether this client should by dynamically placed
or not based on whether TETRIS_DYNAMIC_CLIENT was set or not for the
client. Unfortunately, this doesn't allow a third scenario, namely that
the user doesn't care. Because not setting this variable was equal for
the TETRiS client to static placement of the threads. Hence, now one
has to set TETRIS_MAPPING_TYPE=DYNAMIC. The other possible value is
STATIC which is equivalent to not specifying the variable at all.